vi mode: prevent invalid characters when completing (#579)
When using any type of completion in vi mode, one or more invalid
characters will be output at the end of the completion if the
cursor is on a "wide" character (or if the last character of the
word is wide and the cursor is in the immediately-following
position). This happens because the increment of out for vi mode at
completion.c:280 does not account for character width; out will
therefore be placed in the middle of a wide character rather than
at the beginning of the following character. The fix:

src/cmd/ksh93/edit/edit.{c,h}:
- Add e_savedwidth to store the width of a given character and
  initialize to 0.

src/cmd/ksh93/edit/vi.c:
- When calling the completion, determine the width of the character
  at cur_virt and save it to e_savedwidth.

src/cmd/ksh93/edit/completion.c:
- Using e_savedwidth, increment out by the width of the character
  when in vi mode.

Resolves: https://github.com/ksh93/ksh/issues/571