Fix bug #577 [^]] not working C.UTF8 (#581)
Reproducer (in UTF-8 or ISO locale):

$ ksh -c '[[ "[a] b [c] d" =~ ^\[[^]]+\] ]] && echo ${.sh.match}'
[a] b [c]

Expected output: [a]

The output is already correct in the C locale.

Analysis:

In regconp()-->bra() i.e bracket handling, in the

    if (complicated && mbcoll())
    {

ie (i.e UTF8) when handling ']' in a negative pattern class, i.e.,
[^]...] the ']' is mishandled and then enters wrong elements in the
collate table. It misleads regnexec() into not detecting the ']'.

The original code

    else if (c == ']')
    { ....

The fix

    else if (c == ']' && !(neg && first == pp))
    { ....

If we are in the neg case [^ and if first is '[' and pp is at the
first position don't do this. I then fall back into the normal
processing, effectively entering the ']' on the negate class.

Resolves: https://github.com/ksh93/ksh/issues/577
Co-authored-by: Martijn Dekker <martijn@inlv.org>