[v1.0] Remove SHOPT_RAWONLY
The --viraw option has been forced on by the SHOPT_RAWONLY compile-
time option by default since 93m+ 2002-03-17. This makes sure the
"canonical mode", where user input is echoed by the terminal
without any involvement from the shell, is never used. This was
useful on extremely slow connections in the 1980s but has been
obsolete for decades; it's also incompatible with window resizing.
It's time for the associated code to go. Bit rot has already set
in: ksh crashes in vi_redraw() when compiled with SHOPT_RAWONLY
off, as of 83630f9d, two years ago. No one has complained yet.

This commit removes the SHOPT_RAWONLY compile-time option, and the
obsolete canonical-mode vi code that was only enabled if that
compile-time option was turned off. On the 1.0 branch, a dummy
--viraw option remains for compatibility (forced on, as before).

A note is also added to the man page that the -h/--trackall and
--nolog options (which have had no effect since the early 2000s)
we well as the --viraw option will be removed in ksh 93u+m/1.1.