Mamfiles: replace 'set -' with 'set +v +x' (re: de5bdd12)
This commit replaces the obsolete 'set -' command in the Mamfiles.
In the ancient Bourne shell, 'set -' disables the v and x options,
so is equivalent to 'set +v +x'.

'set -' still works on almost all modern shells, but not yash, and
POSIX considers it "unspecified".

This was already done for iffe in de5bdd12.