build: Work around Apple Xcode linker breakage
The macOS 13 SDK that comes with Xcode.app includes a libmd stub
that fools the linker into thinking the md (message digest) library
is actually installed, though this is not the case. Consequently,
after installing Xcode.app, compiling programs with -lmd
incorrectly succeeds, but executing them fails:

$ echo 'int main(void) { return 0; }' >test.c
$ cc test.c -lmd
$ ./a.out
dyld[63517]: Library not loaded: '/usr/lib/libmd.dylib'
  Referenced from: '/usr/local/src/ksh93/ksh/a.out'
  Reason: tried: '/usr/lib/libmd.dylib' (no such file), '/usr/local/lib/libmd.dylib' (no such file)
Abort

This fools the probes for libmd in lib{cmd,sub}/Mamfile as they
only try to compile the program, not run it. A false positive
result for libmd results, and the build fails at a later (seemingly
unrelated) stage.

Clearly, this is Apple's bug and not ours. As a workaround for
their 93u+ 2012-08-01 build, Apple patched lib{cmd,sum}/Mamfile to
simply remove probing for libmd. But that is not an acceptable
multi-platform solution. A better workaround is almost as simple:
the test program should actually be run as part of the probe.

On Apple systems with this bug, a dyld Abort crash like the above
will now show up in the compile log four times as the test program
fails to run. I'm okay with this.

src/lib/libcmd/Mamfile,
src/lib/libsum/Mamfile:
- When generating cmd.req and sum.req, actually attempt to run the
  test program as part of the link test. This is done by adding
  '&& ./1.${COTEMP}.x' where appropriate.
- Reformat the probe code (originally generated by nmake) for
  legibility and manual maintenance.

Thanks to @tvalenta and @posguy99 for their help finding the cause
of the problem.

Resolves: https://github.com/ksh93/ksh/issues/567
Resolves: https://github.com/ksh93/ksh/issues/596