Fix $"..." translation failure (re: f38494ea)
The referenced commit introduced a failure in $"..." translation as
the strgrpmatch() call in _ast_iconv_name() in iconv.c has a type
mismatch. strgrpmatch() is defined as a call to strngrpmatch() with
an additional STR_INT bit flag passed in 'flags'. But that is only
valid when the pointer to the array passed in the 'sub' argument is
an array of ints.

I found two strgrpmatch() calls in the code where it is wrong for
it to pass STR_INT, as the array used is of type ssize_t. One of
those calls is in iconv.c, causing $"..." translation to fail. The
other is in ccmapid() in ccmapid.c.

All the other strgrpmatch() calls either explicitly and correctly
pass the STR_INT flag for an int array, or pass a null array
pointer, making it irrelevant. So the fix is simple:

src/lib/libast/string/strmatch.c: strgrpmatch():
- Do not pass STR_INT to strngrpmatch().

Thanks to Koichi Nakashima (@ko1nksm) for the report and
reproducer, and for putting me on the trail to the fix.

Resolves: https://github.com/ksh93/ksh/issues/599