Fix bug in nameref loop detection
$ (nameref a=c c=a)
-ksh: typeset: c: invalid self reference
$ (nameref a=b b=c c=a)
[no output! should be same as above]

This is a bug, introduced in ksh 93u 2010-11-22. It introduced a
special '.deleted' node to which name references were set that
refer to variables that are (subsequently) unset. But when that
variable is another name reference to a nonexistent variable, then
self-reference loop detection is defeated.

src/cmd/ksh93/sh/name.c:
- nv_delete():
  - Introduce recognition of the NV_REF bit flag. Do not
    set NullNode (a.k.a. '.deleted') when this is passed.
  - Backport a minor change from ksh 93v-: use dtremove instead of
    dtdelete (the former does a literal name search instead of a
    matching search) and check that a non-null pointer is returned.
- nv_setref(): Pass NV_REF bit flag when calling nv_delete().

Thanks to @phidebian for the report.
Resolves: https://github.com/ksh93/ksh/discussions/574