path_alias => path_settrackedalias (re: 2b3a766d)
This is a code consistency commit that makes the code for setting a
tracked alias analogous to the code for getting one. The
nv_search() call is similarly moved to within the function, and a
consistent check for SH_DEFPATH and SH_XARG is added to it to avoid
a tracked alias ever being created with either of these states on.