fault.h: remove two unused pseudosignal trap entries
The sh.st.trap array (see struct sh_scoped in shell.h) is only used
for the pseudosignal traps 0 (ERR), 1 (KEYBD) and 4 (DEBUG). So 2
and 3 are unused, making that array (declared using SH_DEBUGTRAP
for its size) slightly larger than necessary. There doesn't seem to
be a reason for DEBUG not to be numbered adjacently. These traps
cannot be referred to by number in the trap command, so renumbering
the last one to 2 should be harmless.