sh_trap(): pre-empt an exec optimisation bug (re: 40245e08)
I'm experimenting with new functionality for a built-in command
that invokes a shell function from the built-in using sh_trap(),
and a bug bit me: if the built-in that calls sh_trap() is the last
command executed in a script, then the shell exits when the command
run by sh_trap() completes, not when that built-in completes. Oops.

There is no currently committed code that triggers this bug (and
there won't be on the 1.0 branch), but I'm pre-emptively committing
the fix anyway. On the dev branch, a regression test will be added
later, along with the new code that triggered the bug.

sec/cmd/ksh93/sh/fault.c: sh_trap():
- Set the new sh.st.trapdontexec flag (see referenced commit) to a
  special value 's' if it wasn't already set. The special value is
  used to check whether it is okay to unset it again afterwards;
  this allows commands run by sh_trap() to set traps safely as
  b_trap() resets it to either 1 or 0 when setting/unsetting traps.