Make anchored empty pattern match in ${parameter/pattern/string}
I was not going to backport this fix to the 1.0 branch, but I
reconsidered[*].

@ormaaj wrote:
> This is useful for compatibility and I think it's supposed to
> work anyway.
>
> Working in ksh93 v- and u+m:
>
>     ${x/~(E)^/y}
>
> Non-working:
>
>     ${x/#/y}
>     ${x/#~(K)/y}

The current ksh93 behaviour, though unique, /is/ logical. Unlike an
empty regular expression, an empty shell pattern matches nothing.
The non-working reproducers tell ksh to match an empty glob pattern
at #, the beginning of the string (the ~(K) signifying a shell/glob
pattern, which is also the default). Since empty shell patterns
never match, no substitution is done.

Having said that, mksh, bash, and zsh have all special-cased
${var/#/prefix} and ${var/%/suffix}. I think we should do the same,
because the strictly logical behaviour is not meeting user
expectations and the special-casing is useful, particularly in
vector expansions such as ${@/#/prefix} (which will now prefix each
positional parameter).

This change is a potential corner-case incompatibility, so it is
applied only to the dev branch for the future ksh 93u+m/1.1.

(The ${x/#~(K)/y} case still doesn't work after this commit, but
I'm okay with that; it explicitly circumvents the special-casing
and basically says "yes, I absolutely want to match the string
against an empty glob pattern". So that case is not going to break
user expectations.)

src/cmd/ksh93/sh/macro.c:
- When calling strngrpmatch(3) (in the '%' case, via substring()),
  if pattern is empty, pass an ERE anchor (^ or $ prefixed by ~(E))
  to match the beginning or end of the string, as appropriate.

Resolves: https://github.com/ksh93/ksh/issues/558

________
[*] With the 1.0.5 release around the corner, I am reconsidering
the decision not to have this fix on the 1.0 branch.

On 1.0 and legacy ksh, ${x/#/y} is a no-op; it's equivalent to $x.
There is no use case for that, so there is no good reason why any
legacy script would use it – *except*, of course, if the author
expected it to work. In which case, fixing this in ksh means fixing
that script. Fixing it will therefore likely do more good than
harm.

The existing behaviour may have been technically correct (an empty
glob pattern matches nothing even if anchored), but that is a very
obscure kind of technically correct. Fixing it will align ksh
93u+m/1.0.5 with reasonable user expectations (particularly since
this has been working as expected for many years on mksh, zsh and
bash).

Between that, and the fact that there is no use case for the
original behaviour, fixing it seems like the right thing to do –
particularly as the 1.0 branch has already seen fixes for egregious
design flaws (such as the initial octal zero mess, or matching '.'
and '..' in glob) that are actually more radical than this one.