Fix parsing of nested compound assignments with (( in comsubs
The comsub() function skips over both command substitutions of the
$(...) and ${ ...;} varieties and arithmetic expressions of the
form $((...)). These are all stored in the parse tree in source
code form and parsed at runtime; the skipping is done to read them
and determine the length of the source code text to store in the
parse tree.

When it comes to command substitutions, that skipping is a hack. It
tries to do this using the lexer only and without fully parsing the
command substitution, and is mostly successful, but confusion could
arise due to the complexities of the shell grammar. Hence this bug:

    $ ksh -c 'got=$(typeset -a arr=((a b c) 1); typeset -p arr)'
    ksh: syntax error at line 1: `(' unmatched

There is no syntax error there. The root cause of the bug is that
the double parentheses (( are parsed as the beginning of a ((...))
arithmetic command. As the comsub() hack only does lexical analysis
and not parsing, it has no way of knowing that the (( does not open
an arithmetic command in that particular context.

<future>In the long term, a complete and robust fix will be to
replace this approach for command substitutions and only keep it
for arithmetic expressions. Command substitutions should be fully
parsed like any other code. Their parse subtree, and not their
source code text, should be stored in the parse tree. This would
also make them run more efficiently: a command substitution would
no longer need to be parsed each time it is executed. It will be a
forward and backward incompatible change in the binary format,
which is also used by shcomp, so it will have to wait until a
future release.</future>

For now, the following hack-on-top-of-a-hack will have to do.
Thanks to @atheik for the debugging groundwork that eventually led
to finding this fix.

src/cmd/ksh93/include/shlex.h,
src/cmd/ksh93/sh/lex.c:
- comsub():
  - Whenever we encounter an LPAREN '(', check if the
    previous character was '='. If so, set a new dolparen_eqparen
    flag to the current parenthesis nesting level -- but only if
    it wasn't already set before; we do not need to know about
    nested levels of =( because arithmetic expression cannot occur
    within compound assignments.
  - Whenever we encounter an RPAREN ')', unset the flag if we have
    reached the nesting level of the previously detected =(.
  The above changes cause the dolparen_eqparen variable to always
  be set if we are within =( ... ), regardless of nesting.
- sh_lex(): When encountering '((', do not switch to the ST_NESTED
  state table if the previous character is '=' or when the new flag
  is set. We need both checks to deal correctly with nested
  parentheses, as this code path may encounter =(( before comsub()
  can get around to setting the flag. This change disables the
  recognition of the (( symbol for arithmetic commands within
  compound assignments when these are lexed via the comsub() hack,
  which fixes the bug.

Resolves: https://github.com/ksh93/ksh/issues/269