spawnveg (libast): remove vfork support (re: b810e2d0)
spawnveg(3) too engaged in undefined behaviour according to POSIX
if its vfork fallback version was compiled in. It may have worked
well enough on most systems at the time it was written, but no one
has time to keep verifying that. Plus, all modern systems now have
posix_spawn(3), which is the preferred option for spawnveg. So for
us, this is another opportunity for code cleanup.

This commit does not affect ksh's ability to work on old systems,
it's only that running external commands might now be slightly
slower (and quite possibly more reliable) on old systems without
posix_spawn as they will now use the standard fork(2).