features/standards: fix build on Solaris Studio cc (re: a016a56d)
Solaris Studio cc hates NULL, a.k.a. (void*)0, being used for
function pointers. The POSIX standard says that the conversion of a
void* pointer to a function pointer must work correctly on
conforming implementations[*], but Solaris Studio cc warns when
NULL is used for a function pointer, or in some cases even errors
out (concretely, in src/lib/libast/sfio/sfdisc.c, in the GETDISCF()
and DISCF() macros which use NULL as of the referenced commit).

Instead of finding and fixing all the NULL function pointers, this
commit simply redefines NULL as 0 (instead of (void*)0) for that
particular compiler, as it is always fine in C to use 0 instead of
NULL. Since we're including the standards macros test results from
src/lib/libast/features/standards in everything we compile, that's
the convenient place to do it.

[*] https://pubs.opengroup.org/onlinepubs/9699919799/functions/dlsym.html#tag_16_96_07