mamake: restore old prev behaviour unless MAMAKE_STRICT is set
There is some backward incompatibility with the new `prev` after
all: when building old ksh with the new build system, the ksh and
shcomp binaries aren't generated, though everything else is.
Something somewhere depends on that old dummy rule, it's not clear
to me what.

This reverts `prev` to the exact old behaviour unless the new
`MAMAKE_STRICT` variable is set.