Remove undocumented and buggy label functionality (re: 17f2a55f)
Since <= 1995, ksh93 has been parsing labels ending in ':', called
goto labels in a code comment. In POSIX terms, this is a bug,
because it means command names cannot end in ':' though this is
syntactically perfectly valid in the Bourne/POSIX shell language.

There has never been a goto command that uses these labels, nor is
there any documentation in the manual, in any books, or on the
internet. I've read the source code to figure out what these do.

Only 'break' and 'continue' are sort of capable of using these
labels. In parse.c, the loop nesting levels at which eaech label
occurs are kept track of. There is code in parse.c to translate
'break/continue somelabel' to something like 'break 2' or 'continue
3'. For example, consider the following script:

    for x in 0 1
    do
            echo startloop1
        mylabel:
            echo hi1
            for i in 0 1
            do
                    echo startloop2
                    break mylabel  # executes 'echo startloop1'!
                    echo notreached
            done
            echo reached
    done
    echo end

It outputs:

    startloop1
    hi1
    startloop2
    reached
    startloop1
    hi1
    startloop2
    reached
    end

While this is interesting, there are serious problems with this:

1. Execution resumes at the beginning of the loop level that label
   was detected in, as if 'break' and 'continue' were used with a
   level number. In the script above, 'break mylabel' is translated
   to 'break 1' which is what is actually executed. So execution
   does not necessarily resume anywhere near the label, which is
   against all reasonable expectations.

2. It does not work for arithmetic loops like for((i=0;i<2;i++)).

Problem 2 is easy to fix with a patch to arithfor(), but that does
nothing to remedy problem 1 which makes the whole thing a bad idea.

Both the undocumented label functionality and arithmetic for loops
have been present since 1995 or before. During all those decades,
problem 2 was never fixed. If this functionality ever saw any
actual use, that bug would have been noticed long ago.

So, this is yet another forgotten AT&T experiment that was
abandoned and left to rot in the code base for decades. Its design
is fatally flawed and it causes an incompatibility with other
shells (commands cannot end in ':'). Its removal is long overdue.

This commit is #31 on my list of commits that fix bugs by only
removing code.