Fix completion with histchars and initial tilde (#639)
This commit addresses four completion-related problems.

First: when using completion, any instance of the first histchar
("!" by default) will be left unescaped if it is included in the
result, regardless of whether it had been escaped when the
completion was performed:

    cd \!(tab)
    cd !very\ long\ and\ strange\ directory\ name/

This is perfectly correct if histexpand is turned off. If
histexpand is enabled, however, the resulting command will either
error out or perform an unexpected history expansion. To prevent
this, escape any instance of the first histchar in the result of a
completion if histexpand is enabled.

Second: if the third histchar ("#" by default) is anything other
than #, it will not be escaped at the beginning of the string. This
causes similar unexpected behavior.

Third: if a filename starts with an escaped tilde (~), achieving a
partial completion unescapes the first tilde, making it necessary
to navigate back and fix it. This happens because fmtx only reaches
stakwrite soon enough if an S_EPAT character is in the string, a
category that includes # but not ~. Alter the relevant loop to stop
at ~ as well.

Fourth: when at the start of a quotation, ~ can complete to $HOME
or undergo tilde expansion to an existing user. Send such ~ to fmtx
so it can be escaped properly. This requires temporarily turning
off tilde expansion, for which a new SH_NOTILDEXP state flag is
introduced.

completion.c:
- Escape initial tildes in completion candidates as intended,
  including tildes starting a quotation. In the function fmtx, when
  histexpand is on, add a preceding backslash to the first histchar
  everywhere, and to the third histchar when it appears at the
  beginning of the string.

hexpand.{c,h}:
- Use a separate function to set the three histchars.

Co-authored-by: Martijn Dekker <martijn@inlv.org>