always compile in --globcasedetect option (re: 71934570)
If the functionality is not supported, at least we should still
compile in a dummy option, otherwise 'set --globcasedetect' and
similar will cause profile scripts copied from other systems to
error out on systems where this is not supported.

Of course people can use 'command set --globcasedetect 2>/dev/null'
to avoid erroring out, but non-advanced scripters don't know this
and it doesn't seem reasonable to require it.