don't compile in --globcasedet if explicitly disabled (re: 34e04046)
Other shell options disappear when the corresponding SHOPT_*
compile-time options are set to 0 in SHOPT.sh. So if someone
sets SHOPT_GLOBCASEDET to 0, they probably don't want the
--globcasedetect shell option at all.

We only need the dummy shell option the SHOPT is set to its default
probe (empty) value and the functionality is not detected.

This can be done by checking if the SHOPT_GLOBCASEDET macro is
defined. An explicit disable will define it as 0, but if left to
probe and the probe doesn't detect it, the macro is undefined.