Fix loop variants optimizer vs. discipline function
This is a variant of https://github.com/att/ast/issues/70 that is
still in current 93u+m. Reproducer:

    unset -v var
    for i in 1 2 3 4 5; do
            case ${var+s} in
            ( s )   echo set; unset -v var;;
            ( '' )  echo unset; var.get() { : ; };;
            esac
    done

Actual output:

    unset
    unset
    unset
    unset
    unset

Expected output (and actual output when ksh is recompiled with
SHOPT_OPTIMIZE disabled in SHOPT.sh):

    unset
    set
    unset
    set
    unset

When a .get discipline function definition is used to set the
variable instead of an assignment, the loop invariants optimizer
causes incorrect results.

To fix this, when defining a discipline function, we need to clear
out the saved optimization values for the variable in question.

src/cmd/ksh93/sh/name.c:
- Move the code that clears out a variable's loop invariant
  optimization from nv_unref to its own nv_optimize_clear()
  function so that it can be called from nv_setdisc, which sets the
  discipline function.
- nv_optimize(): Tweak the loop to find the getval/getnum
  discipline. It is used to disable optimization if a getval/genum
  discipline is set. That loop breaks when OPTIMIZE_disc is found,
  but that means it won't find any getval/getnum disciplines set
  after the optimize discipline. Instead, assign the appropriate
  variable (xp) without breaking the loop. ( I have not found any
  bugs that this may fix, but the current code looks incorrect and
  the fix doesn't cause regressions either.)

src/cmd/ksh93/sh/nvdisc.c:
- nv_setdisc(): Call nv_optimize_clear() for the variable in
  question when setting a discipline function. This fixes the bug.

https://github.com/ksh93/ksh/issues/637