tests/functions.sh: Fix regression test (re: 072b0b75)
The addition of 'cd' to the fun directory mysteriously caused
one of the regression tests to fail, though that directory wasn't
in FPATH.

The culprit is the .paths file installed in arch/*/bin. This
contains the line FPATH=../fun, which causes ksh to look for
functions in the fun directory even if FPATH isn't set. The
'typeset -fu cd' command in that tests was not expecting to find
anything, and now does, so the test fails.

src/cmd/ksh93/tests/functions.sh:
- Minimal fix to the regression test. We don't need the old $PATH
  there in the subshell. Removing it avoids that .paths file being
  used. I've verified the regression test continues to fail in the
  same way on ksh 93u+ 2012-08-01 and now again succeeds on 93u+m.