Fix spurious syntax error in 'case' with extended expression
Reproducer:

  $ case fooooo in ~(E)^fo{2,}$) echo ok;; esac
  -ksh: syntax error: `}$' unexpected

Expected output: 'ok'. The ERE (extended regular expression) should
match and not cause a syntax error.

This bug was introduced in ksh 93t 2008-10-31 which has the
following entry in ksh93/RELEASE:

    08-10-24  The lexer now accepts all RE characters for patterns
              prefixed with a ksh ~(...) option expression.

The code added to handle this is run when the ST_NESTED state table
is active (see data/lexstates.c). It uses the S_POP state to handle
the terminating closing parenthesis in 'case' selectors and break
the loop on S_POP when processing a regular expression (TEST_RE) in
'case'. However, the closing brace '}' also has the S_POP state in
that table, though a '}' is a perfectly legitimate character in an
extended pattern. This explains the peculiar syntax error.

src/cmd/ksh93/sh/lex.c: sh_lex(): case S_POP:
- Do not break the loop upon encountering '}'.

Resolves: https://github.com/ksh93/ksh/issues/649