sh_subshell(): make sure SH_TTYWAIT is off (re: 60b3e3a2)
Subshells never wait for tty input because they cannot be
interactive, so the SSH_TTYWAIT shell state (indicating the shell
is waiting for keyboard input on the command line) is as
inappropriate for them as SH_INTERACTIVE is. It should be turned
off whenever entering a subshell of an interactive shell.

The code reverted by the referenced commit accidentally prevented
SH_TTYWAIT from being on in command substitutions invoked from PS1
prompts. Reverting it reintroduced a corner-case problem where
job_unpost() may be inappropriately called on line 485 in jobs.c.
This commit prevents this in a more logical and comprehensive way.