Prevent other ksh93 sessions from interfering with the full editor (#658)
In both line editors, it is possible to run the full editor on an
unexpected command if more than one shell session is running
simultaneously. To reproduce:

  * Open two ksh sessions
  * In one session, enter any command or comment.
  * In the other session, without reaching a new line or using
    history navigation, attempt the full editor command.

The command entered in the first session will appear in the editor
window rather than the contents of the command line in the second
session. This happens because the history line in the second
session is overwritten by the first session. Annoyance aside, the
behavior could also easily result in an unexpected command being
run (if, on seeing unexpected output, the user instinctively quits
the editor without saving, for example).

src/cmd/ksh93/edit/{emacs,vi}.c:
- In case another shell session has used the current history line,
  when on the last history line, use hist_eof and adjust the
  maximum line and current line before calling ed_fulledit. In
  addition, use hist_flush before ed_fulledit if the history line
  is at least as high as histsize.