libast/tm/tvsleep.c: remove assert.h use (re: 6e06e9c5, 4e47f89b)
Including <assert.h> after <ast.h> caused the build to fail on my
museum-grade PowerMac G5 with gcc 3.3. This may be fixed by
including <assert.h> (which is a non-AST system header) before
<ast.h> instead, which should not cause any problems.

However, the assert(3) macro is only used once to verify that the
tv pointer passed to tvsleep(3) is non-NULL. This is non-essential
as the pointer is dereferenced immediately after, and dereferencing
a null pointer reliably triggers a traceable segfault. None of the
other AST code uses assert(3), so let's just get rid instead.

This fixes the build on (at least) the system mentioned above.