Fix 'exit' default status in trap action (re: 092b90da)
Reproducer:

   (trap 'false; exit' EXIT; exit 9); echo $?

Expected output: 9
Actual output: 1

In a trap action, the default exit status passed down by an 'exit'
with no arguments is the exit status of the last command *before*
the trap action.

So, *that* was the point of the sh.oldexit variable removed in
092b90da. BUG_LOOPRET2 was that sh.oldexit was unconditionally used
for the default exit status of 'exit' and 'return' in b_return().

The fix is to restore it, while using it correctly this time (i.e.,
depending on sh.intrap) so that BUG_LOOPRET2 is not introduced.

Thanks to @sckendall2 and @jghub for the bug reports.

Resolves: https://github.com/ksh93/ksh/issues/667
Resolves: https://github.com/ksh93/ksh/issues/687