nvtype.c: fix accidental deletion (re: a04ccf1b)
This was a mistake made while backporting this commit from the dev
branch. The iop variable was deleted as unused in the dev branch
(as we no longer dump function definition source texts there, see
f21ace5c) but is still used in 1.0. The accidental deletion caused
a couple of crashes in tests/types.sh.

Related: https://github.com/ksh93/ksh/issues/699