build system: don't keep rerunning test for -lm (re: 43425b00)
The test for the need to link to libm in libast/Mamfile used the
presence of the astmath.exe target file to determine this. Problem
is, if the file was deleted (and thus -lm is needed), that target
was out of date again, and thus it was rerun on every subsequent
make, along with everything else that depends on it. This occurred
only on systems that were found to require -lm (e.g. FreeBSD).

src/lib/libast/Mamfile:
- Instead of conditionally deleting astmath.exe, write the result
  of the test to astmath.out and use that as the make target.