mamake: drastically simplify prerequisite scanning (re: 6cc2f6a0)
The scan() function scans Mamfiles (originally, nmake Makefiles)
for library prerequisites. When recursively building directories,
it is used via a walk() call to sort the leaf directories in the
order that they need to be built.

The scan() code is designed to scan nmake Makefiles and is very
complicated. But since we eliminated nmake and the Makefiles and
now maintain Mamfiles by hand, we now scan Mamfiles. The Makefile
scnaning code somehow worked on the Mamfiles, more by luck than by
judgment; back in January 2021 I did not have this figured out yet.

Scanning Mamfiles is very simple; this sort of processing is what
they are designed for. All the prerequisite information is in
commands like "bind -lfoo" to declare a dependency on libast. Such
a command causes the leaf directory named libfoo to be built first.

src/cmd/INIT/mamake.c:
- Restore libast getopts documentation about prerequisites, removed
  in the referenced commit (because I lacked understanding).
- scan(): Only process "bind" commands in mamfiles, parsing their
  very simple form. Eliminate all the Makefile-related cruft.