build system: add notrace attribute to mamake
Mamake turns on xtrace for shell actions by default, but some of
our new shell actions in the Mamfiles are quite involved and would
generate too much noise in the log to xtrace. So we turn off xtrace
with 'set +x'. But on all shells other than ksh93, that is still
echoed in the log, which is ugly. A notrace attribute that prevents
'set -x' from being executed in the first place is more elegant.
This commit adds and documents that and makes our Mamfiles use it.

It also adds a virtual rule at the end of ksh93/Mamfile to show the
version of the ksh just built and tells people to 'bin/package use'
to try it out.