mamake: reduce backward compat stubs (re: 856dfb5b)
Archaeological research on the Mamfiles in the ast-open-history
repo shows that the assignment-argument parsing code in the
'silent' and 'ignore' scripts was never used, because no
assignment-argument (e.g.: silent var=value somecommand) was ever
passed. So we don't need to try too hard to be backward compatible
with this.

src/cmd/INIT/mamake.c: run():
- Reduce the 'silent' stub to an empty alias. Any theoretical
  assignment-arguments will be true shell assignments (e.g.:
  'var=value somecommand') instead; the effect should be the same.
- Reduce the 'ignore' prefix to an 'env' command invocation plus
  ':' to ensure a zero exit status; 'env' will process any
  theoretical assignment-arguments. This is not safe for field
  splitting, but then again, this whole system never was.
- Since we're not using them any more, don't bother setting them
  at all in MAMAKE_STRICT mode.