mamake: drop support for nonexistent 'bind +lFOO'
Never in the history of the entire ast-open-archive repo has there
been a 'bind' command in a Mamfile with an argument starting with
'+'. We do not need to be backward compatible with something that
was never used and in fact does not make sense.

Why does it not make sense? In require(), line 1638, any '+l...'
argument is passed directly to the compiler, causing a 'no such
file or directory' as cc does not recognise arguments of that form.