Mamfiles: remove a large amount of redundant cruft
This commit gets rid of roughly 700 prerequisite statements that
refer back to a library (mostly libast) that was already built in
another directory -- those like
	make ${PACKAGE_ast_INCLUDE}/ast.h implicit
...and all its own prerequisites. Those are all redundant, because
mamake sorts the directories in the correct build order by scanning
the 'bind -l...' commands in the Mamfile to determine the
dependency hierarchy (see README-mamake.md for info on that). So,
those files are known to exist.

Even if they weren't (e.g. if you deliberately try to build ksh
without building or updating libast first), those prereqs wouldn't
be all that useful. All they can do is error out if one of the
prerequisites is missing. Now, it will error out at compile time
instead, when the #include fails. Big deal.

This commit also gets rid of some prerequisites marked with the
attributes "implicit dontcare virtual". Those have no effect. They
refer to system headers on certain OSs, not to files we ship.