mamake: fix command name matching
The other day I made a typo in a mamake command: 'exdc' instead of
'exec'. It was treated as 'exec'.

In make(), the four letters of command names are combined into one
'long' value using the KEY() macro, and those values are then
compared in a switch block to match command names efficiently.

The problem is in the KEY() macro. It shifts left 5 bits for each
letter, which is not enough. There is a tolerance for a four-letter
variance, e.c. 'exdc', 'exec', 'exefc' and 'exegc' are the same.

The type used is 'unsigned long' which is guaranteed to be at least
32 bits long, neatly fitting four 8-bit bytes. So there is no
reason not to simply shift 8 bits per letter. This fixes the bug.

Since the Mamfiles were originally generated by AT&T nmake, there
were no command name typos in the originals. Nor are there any in
our current Mamfiles (this fix would have found them). So I don't
have backward compatibility concerns here.