include/ast.h: remove over-optimisation of streq/strneq macros
These macros are used instead of strcmp(3) or strncmp(3) at a few
hundred places in the code. Unlike the latter, the macros return a
simple 1 for true/equal and 0 for false/not equal.

In the 21st century, library functions like str(n)cmp are generally
inlined and/or heavily optimised, so there is no longer a need to
optimise those macros by comparing the first byte separately before
calling str(n)cmp.

This commit also removes some unused EXIT* macros.