ksh: shell quoting: sh_isprint(): fix for non-UTF-8 multibyte
The sh_isprint() function is used in sh_fmtq() (shell quoting) to
figure out if a character is an invisible character, excluding
ASCII space, so that it can replace those with hexadecimal codes.
It was not correct for non-UTF-8 multibyte locales. We don't
support those (yet), but we might as well do it right.

src/cmd/ksh93/sh/string.c: sh_isprint():
- Better optimisation for ASCII in multibyte locale.
- Do not use the UTF-8 fallback when not in a UTF-8 locale.
  The lcinfo function allows us to find this out (we need lc.h).