build system: relink ksh/shcomp if libraries change (re: 4eaf25c8)
When libast.a changed, ksh or shcomp were not getting relinked. To
make that happen, we have to repeat the 'bind -lshell' command,
which acts like 'prev' not just for libshell.a but also for all its
previously built dependencies from other directories.

Similarly, libcmd.a needed an extra 'bind -lsum' to re-extract
libsum.o and re-generated libcmd.a when libsum.a changes.

This commit also removes a few superflous 'prev' commands from some
Mamfiles and updates README-mamake.md with more complete info about
the effect of the 'bind' command.