Fix rare crash in 'print'
On my system, the following intermittently crashes ksh:

	print --???noESC '--???'

Cause: the 'disc' Optdisc_t struct is not fully initialised before
passing it on to optget via opt_info.disc; only two of its four
members were assigned a value. The uninitialised disc.catalog
pointer caused the crash in this case (in optget.c line 1071).
The fix is to memset the struct to zero first.