ed_putchar(): fix remaining potential overflow (re: 9ff7f3db)
Rereading ed_putchar() made it obvious there is still a potential
buffer overflow: there is no bounds check while writing the bytes
of a multibyte character. The bounds check at the end of the
functoin is designed for writing one byte at a time.

src/cmd/ksh93/edit/edit.c:
- Split ed_putchar() into ed_putbyte() and a new ed_putchar(). The
  former is the old non-multibyte aware version. The latter has the
  multibyte stuff, then calls ed_putbyte() for each byte. This may
  cause the buffer to be flushed in the middle of a multibyte
  character, but that should not be a problem as everything should
  be written to the terminal in the correct order in the end.
- If SHOPT_MULTIBYTE is disabled, simply rename ed_putbyte() back
  to ed_putchar() and do not compile the new ed_putchar().

This approach is (obviously) not performance-optimised, but neither
is that necessary: it is only used to output small strings in line
editors that spend 99.999999% of their time waiting for user input.