libast: remove unused LC_XLITERATE locale keyword
LC_XLITERATE was an abortive attempt to standardize transliteration
information in locales. Googling LC_XLITERATE gives very few
results, most of which are arguments of why it was a bad idea
(short version: transliteration is far too complex a problem to
hardcode a method in a locale). See, e.g., "Objection #32" at:
https://www.open-std.org/jtc1/sc22/WG20/docs/n883r-doc-dtr-14652.htm

Searching ast-open-archive shows that there has never been any
AST code that actually used this category, so it's just been
pointlessly taking up a little memory. This commit removes it.

All the other AST locale categories are either POSIX standard or
are supported as extensions by GNU libc, so they are good to have.