math.tab: correct return type for finite()
Trying to build ksh on Android/Termux produced an error regarding
an incompatible library declaration of the finite(3) function. This
is declared in math.tab as a function returning a floating point
value, which is not correct; the non-standardised finite(3) returns
an integer value of 0 or 1 on every system where it exists.
Changing the return type from 'f' to 'i' fixes the build error.