make.probe: rm nostartfiles probe (re: 0c40e7c1, 0b36868c)
The result, stored in $shared → CC.SHARED → ${mam_cc_SHARED}, has
never been used in the Mamfiles, only in the nmake libast/Makefile.

I broke the test for GCC in 0c40e7c1 (an empty grep result was
treated as GCC 7+, which disabled the probe, disabling the probe
for all non-gcc compilers) but it's irrelevant for us as we don't
and won't use nmake.

Given what the -nostartfiles option is for, I don't see how it
makes sense to use it for a userland program like ksh anyway.