build: restore compat with old NetBSD sh (re: f8f58c11)
I'd used expansions of the form ${foo-} for compatibility with a
possible future use of 'set -u'/'set -o nounset'. However, on older
NetBSD sh (as still in use on sdf.org and sdf-eu.org), if foo is
unset, then empty removal is not performed as it should and even an
unquoted ${foo-} yields an empty field. So, to be compatible with
NetBSD sh, we cannot use expansions of that form in contexts where
empty removal is expected.