Fix build on Cygwin (re: aa5e9828, cda1fc25)
On Cygwin, the first referenced commit broke mamake after it got
relinked against libast; fork(2) failed then. Example log output:

| # src/lib/libast/Mamfile: 5007-5024: make mamake
| + cc -fno-strict-aliasing -o mamake mamake.o libast.a
| + set +f
| + rm -rf /usr/local/src/ksh/arch/cygwin.i386/bin/mamake /usr/local/src/ksh/arch/cygwin.i386/bin/mamake.exe
| + cp mamake /usr/local/src/ksh/arch/cygwin.i386/bin/mamake
|
| # ... making src/lib/libsum ...
|       2 [main] mamake 4847 child_copy: data read copy failed, 0x407000..0x4070A0, done 0, windows pid 1324, Win32 error 998
| mamake: *** exit code 139 making lib/libsum
| mamake: *** exit code 1 making all
| package: make failed at Mon Jul 15 17:47:05 BST 2024 in /usr/local/src/ksh/arch/cygwin.i386

Evidently, mamake(2) cannot fork on Cygwin after having been
replaced on the file system (though, with the previous method, it
could still use system(3)).

It works to simply restart the build, as mamake will not need to
be replaced again -- but that is not good enough. I could restore
the deprecated system(3) method for Cygwin, but I don't want to.

bin/package:
- Add a workaround for Cygwin. To allow relinking the binary
  against libast later, add a _bootstrap suffix and create a
  'mamake' symlink to mamake_bootstrap. The symlink will then be
  removed/renamed instead of the mamake_bootstrap binary.

Thanks to Roland Mainz (@gisburn) for the report.
Resolves: https://github.com/ksh93/ksh/issues/758