histexpand: do not process ! after ${
By default, the history expansion character is !, but that
interferes with indirect expansions like ${!varname}, which is a
nuisance on a command line with --histexpand on. So this commit
makes ksh act like bash appears to: make the character pair '${'
escape an immediately following '!', in the manner of a backslash.

Thanks to @posguy99 for flagging this up, even if it was more than
three years ago :-)  See the discussion at:
https://github.com/ksh93/ksh/issues/183#issuecomment-786841942
(note that I'm not so against history expansion now that the
histreedit and histverify options are available; see a5700d39)