libast: cleanup: remove unused pathprobe() function
This function was part of the defunct internal AT&T AST build
system. Grepping the ast-open-archive repo shows it was used by
nmake(1), probe(1), and libpp (pp(3)), the latter being a library
supporting the AST implementation of cpp(1), the C preprocessor.
According to the documentation, the pathprobe function yields the
path to probe(1) information for arbitrary language processors.

There's not going to be any interest in this in the future as we
are going to keep our version of the build system small and
self-contained, so we can get rid.