Fix namespace regression with 'read' (re: 783d3ba4, cc492752)
The referenced commit quietly changed the dtinsert() call in the
low-level variable search function nv_search() to a dtinstall()
call.

Since that commit, 'read' fails in namespace blocks if the variable
never previously existed:

  $ namespace foo { read _DeFiNiTeLy_nOnExIsTeNt_vAr_; }
  -ksh: read: _DeFiNiTeLy_nOnExIsTeNt_vAr_: cannot create

But if the variable first exists is then unset, this bug is not
triggered.

I've experimentally determined that the unexplained change from
dtinsert() to dtinstall() is the cause of the bug.

Making changes without explanation is certainly not my policy or my
habit! A year later, I don't know what I was thinking. Perhaps it
was an experiment that I committed accidentally.

The cdt(3) man page documents:

| For Dtset, Dtrhset or Dtoset, if there is an object in dt
| matching obj, dtinsert() and dtappend() will not insert a new
| object, whereas dtinstall() will remove such a matching object
| before inserting the new object.

This does make sense of the change. By this description, using
dtinstall() to create a node in nv_search() seems more robust.

But, for some reason, dtinstall() returns NULL when a namespace is
active, at least when invoked via ksh's 'read' built-in. This then
triggers a 'read' error: "cannot create".

I've not succeeded in tracing the cause of the returned NULL
pointer so far. The Cdt code is difficult to understand.

It could be that dtinstall() is buggy. It was written during the
93v- beta days and backported to 93u+m in cc492752. But nothing in
the last version of the AT&T repo (2016-01-10) uses it.

This commit reverts the change in question, fixing the bug.