Fix 'typeset -T' overriding existing builtin (re: f43bb498)
Reproducer:

  $ ( (typeset -T echo=(typeset BAD)); echo OK )
  -ksh: echo: panic: internal error

And from that point on, 'echo' is broken and you have to reboot the
shell to get it back, even though the 'typeset -T' was executed in
a subshell. Which makes sense if you follow the referenced commits:
the preliminary dummy declaration built-in is added at parse time.

Analysis: The problem is that sh_addbuilt-in() adds built-ins to
sh.bltin_tree, not dcl_tree. The view from sh.bltin_tree to
dcl_tree doex cause the built-in to be added to the latter, but
only if a built-in by that name does not exist in sh.bltin_tree.
If one does, it gets overwritten.

src/cmd/ksh93/sh/parse.c: check_typedef():
- Temporarily set sh.bltin_tree to dcl_tree while calling
  sh_addbuiltin, avoiding the possibility of it finding and
  overwriting existing built-ins.
- Rewrite the explanation about the dcl_tree hack in the comments
  to make it clearer (I hope).