regress: pty: add kludge to tolerate buffer underflow (re: c43086b8)
Urgh. So, the 'r < bp->buf' buffer underflow problem is system-
dependent: one pty.sh test that passed fine on my Mac fails on the
Github runner.

I've tried various ways of detecting and correcting the underflow,
but they all break pty and make (some of) the tests fail.

I can't for the life of me figure out how it underflows, but in the
cases I've observed so far, it's underflowed by exactly 9 bytes.

src/cmd/builtin/pty.c:
- Allocate/initialise BUFUNDERFLOW extra bytes before every buffer,
  where BUFUNDERFLOW is currently 128 bytes. Given that I've only
  observed 9 bytes of underflow so far, that seems fairly generous.

]This is quite a pathetic kludge and I hope it can be fixed properly
someday, but for now, all the tests pass under ASan (including the
one disabled in the referenced commit, which is now reenabled).