build: mamake: consolidate 'prev' code into function
There are six places where mamake() had a procedure to propagate
the timestamp from the child rule to the current rule in the manner
similar to the 'prev' command -- of course in slightly different
variants. This could make it easy to introduce bugs in future.

So we now add a propagate() function that centralises this logic.

This means we're no longer using the return values of bindfile()
and run(), which was redundant anyway as the same info was in
(rule)->time, so those are now void functions.