libast/sfio/sfrd(): fix unpredictable misbehaviour (re: eab10941)
I did a goof in sfrd() and left the 'r' variable uninitialised if
f->data is NULL. This commit makes sure r gets initialised to zero
in that case.

This is a very central function, so this was causing either severe
regressions in all sorts of places, or no regressions at all.
That's undefined behaviour for ya...