build: mamake: rm include <assert.h> (re: 9984004c)
On 32-bit Linux systems, as soon as mamake was relinked against
libast, the libast version of fseeko(3) was displaying some really
weird undefined behaviour, breaking the 'loop' command. The culprit
was the <assert.h> include. The problem went away as soon as that
was removed.

assert.h was included before libast to avoid a compilation failure
on some other systems, but including it before libast evidently
breaks libast -- it looks like it may pull in parts of Stdio that
libast (which reimplements Stdio in terms of Sfio) does not have a
chance to override, so the two get mixed in breaking ways.

I'd never seen assert.h combined with libast code before I used it
in mamake; now I know why :-/

src/cmd/INIT/mamake.c:
- Remove the <assert.h> include.
- Define assert(c) as a simplistic 'if (!(c)) abort()'. There won't
  be a message if it goes wrong, but that's no big deal; we'll just
  have to follow the stack trace.