build: mamake: fix chaotic output for parallel builds (re: 9984004c)
I said I'd fix it, so here it is. By default, when working in
parallel, mamake now tells the shell to write the output for each
shell action into a temp file. When the action is finished, mamake
writes the trace header, dumps that file to the log, and deletes
it. Result: the log is now quite readable and actually usable for
debugging purposes. Individual jobs will appear somewhat out of
order, but not too much, because it's constrained by dependencies.

Since there is something oddly satisfying about watching the old
chaotic log output flash by (it gives you more of a sense of just
how fast it's working), I've added a `-c` option (for "chaos") that
disables this mechanism and reverts to the previous behaviour.

Main changes (smaller tweaks not listed -- they're obvious enough):

src/cmd/INIT/mamake.c:
- Never run rules marked 'virtual' in parallel. This includes the
  rules that recursively invoke mamake, which would mess up the
  mechanism described below.
- Rule_t: add members for endline (needed for delayed printing of
  the trace header) and the temp log file's name.
- print_nice_hdr(): Added. This consolidates the header-printing
  from make() and recurse().
- reap(): When a background job finishes, call print_nice_hdr(),
  then dump and remove the job's cached output file, if any.
- execute(): Make the shell redirect parallel job output to a temp
  file. Since we don't have a PID before forking, we use a simple
  increasing serial number. Since this temp file will be written to
  the build directory and not a global temp directory, it's safe
  enough this way.