package(1): do not bother with cc wrapper script (re: 5e6c4219)
There were still problems with the CC consistency check this way as
well, e.g., on systems where gcc is installed as 'gcc' and no 'cc'
command is available, the build would only work once, then fail on
rebuild as CC changed. <sigh>

I've never really understood why that cc wrapper script is
installed on PATH anyway. It only hands off control to the command
indicated by the original value of CC. We might as well not change
CC to begin with. So this simply gets rid of the wrapper script,
solving the problem in question.