Fix crash in get/getn disciplines; document getn (re: 2a43cbc3)
Reproducer: see the test added in tests/variables.sh.

Analysis: The global 'struct blocked *blist' variable in nvdisc.c,
which is the starting point for all the bp pointers, gets
invalidated when the arithmetic subsystem throws an error in the
reproducer. Which is obvious, because the struct blocked items in
the list are all local/automatic variables in the assign() and
lookup() functions, which no longer exist after a longjmp.

The root problem is that an error is thrown during the nv_getn()
call, so that block_done() is not getting called after the longjmp,
so the global blist variable is not reset, causing a probable crash
on any subsequent use of a discipline function.

src/cmd/ksh93/sh/nvdisc.c: lookup():
- Delay calling nv_getv or nv_getn until after restoring all
  state, including after calling block_done().

src/cmd/ksh93/sh.1:
- Finally add documentation for the getn discipline.

Resolves: https://github.com/ksh93/ksh/issues/435