[v1.0] Fix SHOPT_EDPREDICT build (re: a45a0ebd, 9ff7f3db, 5b0f2350)
The SHOPT_EDPREDICT compile-time option has been removed from the
dev branch (see #233) but kept on the 1.0 branch. Building with
this option enabled in SHOPT.sh was subsequently broken by the
three referenced commits as changes were cherry-picked from the dev
branch, which of course did not include the necessary code changes
to the SHOPT_EDPREDICT code. This commit backports those changes.

Thanks to @k2662 for the report.
Resolves: https://github.com/ksh93/ksh/issues/786