Add forking workaround for block stdout redir (re: 2d4a7875)
Reproducer:

    ok=$({ true >&2; } >&2; echo OK); echo $ok

This should output "OK" but currently outputs nothing.

Looks like the sh_subfork() workaround must also be implemented for
the TSETIO case in sh_exec(), as that is where redirections for
blocks and compound commands are handled. This avoids inconsistent
behaviour that happens if sh_subfork() is called later for the
inside redirection.

src/cmd/ksh93/sh/xec.c: sh_exec(): case TSETIO:
- Also fork if standard output redirections attached to blocks are
  executed within a non-subshare command substitution.

Thanks to Vincent Mihalkovič (@vmihalko) for the report.
Resolves: https://github.com/ksh93/ksh/issues/784