mamake: fix UB in legacy mode 'prev' command (re: 07a4fc67)
Building old commits that use legacy mode Mamfiles was failing due
to a bug in the 'prev' command.

In mamake.c on line 2392, the pointer returned by the legacy dummt
'rule' call must be assigned to q. It wasn't, so the update_allprev
call on line 2422 had undefined behaviour as it passes q, which is
dereferenced on line 2010.