mamake: unified function for including files (re: f58153d4)
Mamfiles currently include other files with MAM code upon 'setv CC'
to read the compiler probe results from lib/probe/C/mam/HASHEDPATH,
and upon 'bind -lfoo' to read the dependency rules for libfoo from
lib/mam/foo.

To ensure current and future consistency, this commit consolidates
the include functionality in an include() function that also saves
and restores the automatic variables.

This functionality is now also exposed to the MAM language in the
form of a new 'incl' command. It's currently unused, but may come
in handy one day, and it's so trivial it seems silly not to add it.