cp,mv,ln pathbound builtins: properly implement -v/--verbose
The way the -v/--verbose functionality is implemented in cp.c
(which also provides mv and ln) does not seem useful at all, and is
not what users expect. Only when an existing file is about to be
overwritten will -v print the name of that file. Which means that -v
does nothing most of the time.

This commit makes the -v option act exactly like the BSD/macOS
version of these commands: a -v line containing both the source and
the destination (separated by '=>' for creating hardlinks and '->'
for everything else) is printed after each successful operation.
For the 'ln' command, source and destination are reversed.