nvalue sanitisation project part 1: ${.sh.level}
The Namval_t name-value nodes (used for shell variables, functions,
and many other things) use a member 'nvalue' of type 'union Value'
to store the values; see include/name.h, line 32 and on.

These values are usually pointers to allocated memory blocks, but
not always: for certain short int types (attribute bits NV_INT16
and NV_UINT16) the values are stored directly via the 'sp' union
member. This was reduced in d4adc8fc where 'typeset -si' values
began to be stored in small allocated memory blocks (NV_INT16P and
NV_UINT16P), but ${.sh.level} and enum values are still stored
directly in nvalue instead of via pointers.

The problem is, the code contains many points at which the value
pointers are accessed directly, e.g. to check if they are non-null
(e.g., if(np->nvalue.cp) ...). Unless the type of the node is
known, all these checks need to be preceded by a check if the value
is expected to be a pointer, otherwise undefined behaviour may
result. Hence the nv_isnonptr() macro introduced in de462918.
However, this is not yet applied consistently, and in any case this
situation makes it very easy to introduce bugs in future.

So this commit begins a project to sanitise and demystify Namval_t
value handling. To begin with, all the non-pointer Namval_t values
should be changed to pointers, to avoid this class of potential
undefined behaviour. The efficiency reduction should be negligible
while the increase in code robustness is significant.

When all the 'union Value' members are pointers, it should be
fairly easy to get rid of 'union Value' and have a simple void
*nvalue pointer that is used for all the values. The code will be
changed to use typecasts where needed to dereference the nvalue
pointer. This should make the code a little more accessible to
those trying to understand it.

This commit starts it off by fixing ${.sh.level}; instead of
storing its value directly in SH_LEVELNOD->nvalue.s, it now uses
the SH_LEVELNOD->nvalue.sp pointer which is initialised to point to
a new global sh.level variable. This is done by changing its
attribute bits to NV_INT16P in data/variables.c and initialising
the nvalue.sp pointer to point to sh.level in init.c.

Since we're here, this also removes unused 'union Value' members.